feat: sync docs (by copying them)
This implements a different approach to syncing the docs. Rather than creating a toolchain that requires git sync or similar on build and keeps things transient, this approach implements a script that is invoked separately (and can be part of a github action for example) that mostly copies the documentation from the target repositories over into the omnisite, while also invoking scripts for creating the generated reference docs. The result is saved to disk, which has a few advantages:
- build is massively simpler: no dependencies on flaky git or other remote sources, as everything is already there
- getting started is easier, no need to sync first, just checkout the repo and go. This is especially helpful if you're working on the site itself, not the content
- translations will be easier: just don't invoke the sync script, instead modify the directories by hand. Keeping up to date with the source will also be easier because you can git-diff changes to docs

At the same time, this lays the foundation for versioning; kit and svelte docs are both inside a v02/v05 directory (leading 0 to not muss up ordering once we're in double digits)