feat: docs versioning
WIP of having versioned docs. The idea is that we're doing Git clones of the target repositories into our gitignored folder, then copying over the docs / generating the types as needed. A version selector at the top of the side nav then allows you to switch versions.

That all is straightforward, but it reveals a problem: The structure of how we do docs has changed, and the old Svelte docs don't follow that. We either need to adjust them, or add logic to work with the old format, too.